requirements.txt
boto3==1.17.82
botocore==1.20.82
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.61.0
urllib3==1.26.5
